Created a Constant file for all status codes and response messages

+++ src/controller/followController.js
@@ -1,11 +1,24 @@
 const Follow = require('../database/models/followModel');
+const User = require('../database/models/userModel');
+const { StatusCodes, ResponseMessages } = require('../constants/repsonseConstants');
 
 exports.follow = async (req, res) => {
   try {
     const userfollowed = await Follow.findOne({ user: req.user.id, following: req.params.id });
     if (userfollowed) {
-      return res.status(409).json({
-        message: 'You already follow this user!',
+      return res.status(StatusCodes.CONFLICT).json({
+        message: ResponseMessages.ALREADY_FOLLOW,
+      });
+    }
+    if (req.user.id === req.params.id) {
+      return res.status(StatusCodes.BAD_REQUEST).json({
+        message: ResponseMessages.SELF_FOLLOW,
+      });
+    }
+    const user = await User.findById({ _id: req.params.id });
+    if (!user) {
+      return res.status(StatusCodes.NOT_FOUND).json({
+        message: ResponseMessages.NO_USER,
       });
     }
     const follow = await Follow.create({
@@ -13,13 +26,13 @@ exports.follow = async (req, res) => {
       following: req.params.id,
 
     });
-    return res.status(201).json({
-      message: 'Successfully Followed User',
+    return res.status(StatusCodes.CREATED).json({
+      message: ResponseMessages.SUCCESS,
       data: follow,
     });
   } catch (error) {
-    return res.status(500).json({
-      message: 'Could not follow User',
+    return res.status(StatusCodes.SERVER_ERROR).json({
+      message: ResponseMessages.FAILURE,
       error,
     });
   }
@@ -27,22 +40,28 @@ exports.follow = async (req, res) => {
 
 exports.unfollow = async (req, res) => {
   try {
+    const user = await User.findById({ _id: req.params.id });
+    if (!user) {
+      return res.status(StatusCodes.NOT_FOUND).json({
+        message: ResponseMessages.NO_USER,
+      });
+    }
     const following = await Follow.findOneAndRemove({
       user: req.user.id,
       following: req.params.id,
     });
     if (!following) {
-      return res.status(404).json({
-        message: 'You dont follow this user',
+      return res.status(StatusCodes.BAD_REQUEST).json({
+        message: ResponseMessages.DOESNT_FOLLOW,
       });
     }
-    return res.status(200).json({
-      message: 'Successfully unfollowed user',
+    return res.status(StatusCodes.OK).json({
+      message: ResponseMessages.SUCCESS,
       data: following,
     });
   } catch (error) {
-    return res.status(500).json({
-      message: 'Could not unfollow user',
+    return res.status(StatusCodes.SERVER_ERROR).json({
+      message: ResponseMessages.FAILURE,
       error,
     });
   }
@@ -50,19 +69,25 @@ exports.unfollow = async (req, res) => {
 
 exports.seefollowing = async (req, res) => {
   try {
+    const user = await User.findById({ _id: req.params.id });
+    if (!user) {
+      return res.status(StatusCodes.NOT_FOUND).json({
+        message: ResponseMessages.NO_USER,
+      });
+    }
     const following = await Follow.find({ user: req.params.id })
       .populate({
         path: 'following',
         select: '-password',
         model: 'User',
       });
-    return res.status(200).json({
-      message: 'This user is following',
+    return res.status(StatusCodes.OK).json({
+      message: ResponseMessages.SUCCESS,
       data: following.map((follow) => follow.following),
     });
   } catch (error) {
-    return res.status(500).json({
-      message: 'Could not fetch followed users',
+    return res.status(StatusCodes.SERVER_ERROR).json({
+      message: ResponseMessages.FAILURE,
       error,
     });
   }
@@ -70,19 +95,25 @@ exports.seefollowing = async (req, res) => {
 
 exports.seefollowers = async (req, res) => {
   try {
+    const user = await User.findById({ _id: req.params.id });
+    if (!user) {
+      return res.status(StatusCodes.NOT_FOUND).json({
+        message: ResponseMessages.NO_USER,
+      });
+    }
     const followers = await Follow.find({ following: req.params.id })
       .populate({
         path: 'user',
         select: '-password',
         model: 'User',
       });
-    return res.status(200).json({
-      message: 'User is followed by:',
+    return res.status(StatusCodes.OK).json({
+      message: ResponseMessages.SUCCESS,
       data: followers.map((follow) => follow.user),
     });
   } catch (error) {
-    return res.status(500).json({
-      message: 'Could not fetch followers',
+    return res.status(StatusCodes.SERVER_ERROR).json({
+      message: ResponseMessages.FAILURE,
       error,
     });
   }
@@ -96,13 +127,13 @@ exports.userfollowing = async (req, res) => {
         select: '-password',
         model: 'User',
       });
-    return res.status(200).json({
-      message: 'This user is following',
+    return res.status(StatusCodes.OK).json({
+      message: ResponseMessages.SUCCESS,
       data: following.map((follow) => follow.following),
     });
   } catch (error) {
-    return res.status(500).json({
-      message: 'Could not fetch followed users',
+    return res.status(StatusCodes.SERVER_ERROR).json({
+      message: ResponseMessages.FAILURE,
       error,
     });
   }
@@ -116,13 +147,13 @@ exports.userfollowers = async (req, res) => {
         select: '-password',
         model: 'User',
       });
-    return res.status(200).json({
-      message: 'User is followed by:',
+    return res.status(StatusCodes.OK).json({
+      message: ResponseMessages.SUCCESS,
       data: followers.map((follow) => follow.user),
     });
   } catch (error) {
-    return res.status(500).json({
-      message: 'Could not fetch followers',
+    return res.status(StatusCodes.SERVER_ERROR).json({
+      message: ResponseMessages.FAILURE,
       error,
     });
   }
